requirements.txt
click==6.6
decorator==4.0.10
EVEPrest==1.3.5
Flask==0.11.1
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
future==0.15.2
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
praw==3.5.0
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.14
update-checker==0.11
Werkzeug==0.11.10
git+https://github.com/ntt/eveapi
